requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
python-multipart==0.0.9
jinja2==3.1.4
pydantic==2.8.2
pdfminer.six==20240706
pypdfium2==4.30.0
faiss-cpu==1.8.0.post1
numpy==1.26.4
pandas==2.2.2
sentence-transformers==3.0.1
openai==1.44.0
sqlalchemy==2.0.32
python-dotenv==1.0.1